+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### v0.0.3
+
+- Added method `addSourceToCustomer`
+- Added method `deleteSourceFromCustomer`
+
 ### v0.0.2
 
 - Updating documentation